refactor and separating styles from components in other files
