+++ src/components/Cards.jsx
@@ -1,14 +1,5 @@
 import Card from "./Card";
-import styled from "styled-components";
-import { themeGet } from "@styled-system/theme-get";
-
-const Pokedex = styled.div`
-  display: grid;
-  width: 100%;
-  height: 100%;
-  grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
-  gap: ${themeGet("space.sm")};
-`;
+import { Pokedex } from "./Cards.styles";
 
 const Cards = ({ pokemons, isLoading }) => {
   return (
